DNS cache improvements
1. cache based on requested value and not the returned value:
The current behavior of amp's dns cache is to store based on the key returned
from the DNS server. So if we're making a request for an AAAA or DNAME record
and get back a CNAME, it would cache under the CNAME key instead of the records
requested. That means subsequent requests for AAAA or DNAME would have to hit
the DNS server again

2. Load a memcached dump into the in-memory cache so it can live for longer and
be shared across requests and nodes